# Use TextDecoder for improved save file load performance